+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [4.3.1] - 2024-08-16
+- Set simplejson to 3.13.2 for Python 2.
+- Set pymongo to 4.6.3 for Python 3.
+- Updated python-lib to 3.0.4
+
+### Added
+- mongo_libs.create_security_config:  Create security configuration object from a configuration file or mongo instance.
+- mongo_libs.data_out: Outputs the data in a variety of formats and media.
+
+### Changed
+- mongo_libs.create_instance, mongo_libs.crt_coll_inst:  Replaced section of code with call to create_security_config.
+
+
 ## [4.3.0] - 2024-04-09
 - Added TLS connection capability to the Mongo classes and libraries.
 - Set pymongo to 3.12.3 for Python 2 and Python 3.